Update documentation for setting default values
Using `commit_db_transaction` can be problematic when the rest of your code causes the raising of an exception. In that case the migration will not be seen as complete by Rails, so you cannot rollback its changes.

So this could be problematic:

```ruby
#...create columns
commit_db_transactions

#... some code that runs fine in development but causes an exception in production
```

We are now stuck in a state where we can't migrate again (after fixing some records that caused the exception manually for example) in production because as soon as we try to run migrate the columns already exist and raise a new error. This is what we were trying to avoid by using the gem in the first place.

Splitting the process up in 2 commits however fixes this issue. The first creates the columns and the second sets the default and performs the backfilling. This way if there is an exception we can perform some manual fixes, etc and then migrate again. Now only the backfill is run.